Dashboard
In the analysis of survey completion rates, it is observed that the average completion rate during the period is 75%, with variations ranging from 67% to 82% in different months. The high completion rate indicates that paid customers actively engage with the platform and its product. It is noted that there is a certain seasonality in the data, with odd months having more new registered users compared to even months. This is attributed to a bimonthly marketing campaign during odd months, offering annual subscriptions at a discounted price, resulting in an influx of new subscribers. However, not all new subscribers immediately engage with the platform, leading to lower survey completion rates in these months.

The data suggests that new subscribers may require additional time to familiarize themselves with the platform and its features before completing the survey. This is evident in a gradual decline in completion rates over odd months, indicating that early registrants are more likely to complete the survey.

Analyzing the student profile reveals that most onboarded students are new to the data science learning world and are interested in data analysis, visualization, and improving their programming skills. The primary acquisition channel bringing in students is YouTube, where tutorials and career-related video content are frequently posted. The largest group completing the survey is from the United States, followed closely by India, collectively representing nearly half of all participating students. This aligns with the program being in English and the high populations of these countries.

Two actionable insights emerge from the data:

1. **Content Curation:** It is recommended to align the platform's content with students' learning objectives, offering introductory and advanced data science courses. This includes courses on data analysis using various tools (e.g., Python and SQL) and data visualization software (e.g., Tableau and PowerBI). A natural progression and learning path into the data science realm should be provided, incorporating more advanced courses in Machine Learning and projects for practical application.

2. **Channel of Acquisition:** Given that YouTube is the primary acquisition channel, a strategy is proposed to upload introductory data science videos and tutorials from the platform. This can help reach a larger audience interested in data science and analysis. Additionally, paid YouTube videos can extend outreach across the US and India, reaching a broader audience of data science beginners.

These insights serve as key observations for the content creation and marketing teams, guiding the development of the platform to meet the needs of its students. Further analysis is encouraged to generate additional insights and proposals for an effective strategy.